submit now updates, wip
