+++ app/constants/statuses.js
@@ -1,0 +1,4 @@
+module.exports = {
+  IN_PROGRESS: 'In progress',
+  COMPLETED: 'Completed'
+}
