+++ app/inbound/index.js
@@ -1,9 +1,7 @@
-const processPaymentRequest = require('./process-payment-request')
 const processProcessingPaymentRequest = require('./process-processing-payment-request')
 const processSubmitPaymentRequest = require('./process-submit-payment-request')
 
 module.exports = {
-  processPaymentRequest,
   processProcessingPaymentRequest,
   processSubmitPaymentRequest
 }
