+++ app/inbound/update-invoice-lines.js
@@ -1,0 +1,13 @@
+const db = require('../data')
+
+const updateInvoiceLines = async (invoiceLines, paymentRequestId, transaction) => {
+  for (const invoiceLine of invoiceLines) {
+    delete invoiceLine.invoiceLineId
+    invoiceLine.fundingCode = invoiceLine.schemeCode
+    await db.invoiceLine.update({ invoiceLine },
+      { where: { paymentRequestId } },
+      { transaction })
+  }
+}
+
+module.exports = updateInvoiceLines
