+++ app/inbound/update-payment-request.js
@@ -1,0 +1,16 @@
+const db = require('../data')
+
+const updatePaymentRequest = async (paymentRequestId, paymentRequest, transaction) => {
+  await db.paymentRequest.update({
+    ...paymentRequest,
+    received: new Date()
+  },
+  { where: { paymentRequestId } },
+  { transaction })
+}
+
+const updateAndReturnPaymentRequest = async (paymentRequestId, paymentRequest, transaction) => {
+  return updatePaymentRequest(paymentRequestId, paymentRequest, transaction)
+}
+
+module.exports = { updateAndReturnPaymentRequest, updatePaymentRequest }
